Перед разбиением на функции processing files

+++ string_functions.h
@@ -7,7 +7,7 @@ void StrswapGovno(char* temp, char* str_first, char* str_second);
 void StrswapBykva(char* str_first, char* str_second);
 void StrswapPachka(char* str_first, char* str_second);
 
-int MyFputs(const char* string, FILE* stream);
+size_t MyFputs(const char* string, FILE* stream);
 int CountLines(char* text_buffer);
